# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Private Moments API
  version: 0.1.0
  description: API contract for the Mac server used by the iOS app and admin UI.
servers:
  - url: http://127.0.0.1:3210
security:
  - bearerAuth: []
paths:
...
  /api/v1/media/upload:
    post:
      summary: Upload a media file for a post.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - mediaId
                - postId
                - variant
                - file
              properties:
                mediaId:
                  type: string
                  format: uuid
                postId:
                  type: string
                  format: uuid
                variant:
                  type: string
                  enum:
                    - compressed
                    - original
                    - thumbnail
                kind:
                  type: string
                  enum:
                    - image
                    - video
                    - audio
                  default: image
                mimeType:
                  type: string
                durationSeconds:
                  type: number
                transcriptionText:
                  type: string
                  description: Legacy optional transcript metadata for older clients; new iOS uploads omit it.
                originalPreserved:
                  type: boolean
                  default: false
                sortOrder:
                  type: integer
                  default: 0
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - media
                properties:
                  media:
                    type: object
                    required:
                      - id
                      - postId
                      - variant
                      - status
                      - path
                      - sizeBytes
                      - checksum
                    properties:
                      id:
                        type: string
                        format: uuid
                      postId:
                        type: string
                        format: uuid
                      variant:
                        type: string
                      status:
                        type: string
                      path:
                        type: string
                      sizeBytes:
                        type: integer
                      checksum:
                        type: string
  /api/v1/media/batch-download:
    post:
      summary: Download multiple media variants as base64 JSON payloads.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mediaIds
              properties:
                mediaIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                variant:
                  type: string
                  default: thumbnail
                  enum:
                    - compressed
                    - original
                    - thumbnail
      responses:
        "200":
          description: Downloadable media payloads. Missing or deleted media are skipped.
          content:
            application/json:
              schema:
                type: object
                required:
                  - media
                properties:
                  media:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaBatchDownloadItem"
  /api/v1/media/{mediaId}:
    get:
      summary: Download a media variant.
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: variant
          in: query
          schema:
            type: string
            default: compressed
            enum:
              - compressed
              - original
              - thumbnail
      responses:
        "200":
          description: Media bytes.
          content:
            image/*:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    MediaBatchDownloadItem:
      type: object
      required:
        - id
        - variant
        - contentType
        - fileName
        - base64
      properties:
        id:
          type: string
          format: uuid
        variant:
          type: string
        contentType:
          type: string
        fileName:
          type: string
        base64:
          type: string
...